fix(extension): move userIdServiceProperties to config file
