+++ apps/browser-extension-wallet/src/providers/AnalyticsProvider/getUserIdService.ts
@@ -1,6 +1,6 @@
 import { consumeRemoteApi } from '@cardano-sdk/web-extension';
+import { userIdServiceProperties } from '@lib/scripts/background/config';
 import { USER_ID_SERVICE_BASE_CHANNEL, UserIdService } from '@lib/scripts/types';
-import { userIdServiceProperties } from '@lib/scripts/background/services';
 import { runtime } from 'webextension-polyfill';
 
 export const getUserIdService = (): UserIdService =>
